Move default project function, replace .get with []
